refactor: move ts plugin / ts program project logic into vue-tsc and typescript-vue-plugin
